Fix variant unsigned type generation issues
Originally TypeName for mutators and accessors for int8, int16, uint8 were byte or short. Changed those types to int.

+++ src/main/java/org/reaktivity/nukleus/maven/plugin/internal/ast/visit/ScopeVisitor.java
@@ -54,7 +54,7 @@ public final class ScopeVisitor extends AstNode.Visitor<Collection<TypeSpecGener
     static
     {
         Map<AstType, AstType> uintMappings = new HashMap<>();
-        uintMappings.put(AstType.UINT8, AstType.INT16);
+        uintMappings.put(AstType.UINT8, AstType.INT32);
         uintMappings.put(AstType.UINT16, AstType.INT32);
         uintMappings.put(AstType.UINT32, AstType.INT64);
         uintMappings.put(AstType.UINT64, AstType.INT64);
