+++ src/main/java/org/reaktivity/nukleus/maven/plugin/internal/ast/visit/VariantVisitor.java
@@ -45,7 +45,7 @@ public final class VariantVisitor extends AstNode.Visitor<Collection<TypeSpecGen
     static
     {
         Map<AstType, AstType> uintMappings = new HashMap<>();
-        uintMappings.put(AstType.UINT8, AstType.INT16);
+        uintMappings.put(AstType.UINT8, AstType.INT32);
         uintMappings.put(AstType.UINT16, AstType.INT32);
         uintMappings.put(AstType.UINT32, AstType.INT64);
         uintMappings.put(AstType.UINT64, AstType.INT64);
